refactor(cli): Isolate ConfigArgs
